pyproject.toml
[project]
name = "discordvoicerotator"
version = "0.1.0"
description = "A production-ready Discord bot for rotating members between configured voice channels."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.4.0,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
